cloud/requirements.txt
fastapi
maturin
requests
uvicorn[standard]
pycryptodome
